+++ installer/src/surface_cdms/wx_config/core/context_processors.py
@@ -1,8 +1,7 @@
 from django.conf import settings
+from surface_cdms.version import get_surface_version, normalize_surface_version
 
 def get_installer_version(req):
     return {
-        'APP_VERSION': settings.APP_VERSION,
-        'APP_VERSION_STAGE': settings.APP_VERSION_STAGE,
-        'APP_VERSION_LABEL': settings.APP_VERSION_LABEL
+        'APP_VERSION_LABEL': normalize_surface_version(get_surface_version())
     }